Implement PriceController and update PriceService to return Price objects

+++ price-service/src/main/java/com/test/price_service/application/dtos/PriceResponse.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record PriceResponse(
         Long productId,
         Long brandId,
-        Long priceList,
+        Integer priceList,
         String startDate,
         String endDate,
         String finalPrice
